refactor(web): rename chat-context.tsx to chat-store.ts
Align file name with its content — the module is now a Zustand
store, not a React Context. Follows the existing *-store.ts
convention (sse-connected-store.ts). Extension changed from .tsx
to .ts since the file no longer contains JSX.
